Create a base class for unit tests
This is a subclass of unittest.TestCase (or unittest2.TestCase for
Python 2). The main piece of functionality this provides is creating a
`self.bindings`, which behaves exactly as though you had imported a
Python module named `bindings`. The reason for doing this instead of
trying to simply import the bindings file is that this eliminates any
potential issues which could arise from waiting for a disk write and
then a disk read. This code never even touches the disk, and is stored
entirely in memory.