WIP: Convert tests to use unittest (#163)
* Create a base class for unit tests

This is a subclass of unittest.TestCase (or unittest2.TestCase for
Python 2). The main piece of functionality this provides is creating a
`self.bindings`, which behaves exactly as though you had imported a
Python module named `bindings`. The reason for doing this instead of
trying to simply import the bindings file is that this eliminates any
potential issues which could arise from waiting for a disk write and
then a disk read. This code never even touches the disk, and is stored
entirely in memory.

* Converted the typedef tests to use the base TestCase class

* Improve the file path detection to be...readable

* Switch over to using unittest to run the tests

* WIP for moving tests over to using unittest

* Set up the rails for using 'python setup.py test'

* Convert the binary tests to use unittest

* Convert the boolean-empty tests to use unittest

* Set up for test coverage tracking

* Convert the choice tests to use unittest

* Move back to the old test runner for the time being

* Revert these tests since they were only partially transitioned

* Update for a more appropriate test name

* Fix some silliness in the bitarray usage

* Make the flags passed to pyang configurable