fix a identityref field namespace missing issue. the issue is: when load identityref field to Yang data tree with module name as the namespace, when we dump the Yang data to ietf json, the namespace will lost.
the root cause is: when we generate python code from Yang module, the
restriction_arg for the field doesn't include module name as the prefix.

the fix is: add source_module into the imported_prefixes, so pyangbind can
generate also generate the module name as the namespace for identityref field.